requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
moomoo_api==9.6.5608
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==3.20.3
pycryptodome==3.23.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
seaborn==0.13.2
simplejson==3.20.2
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
